% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{calculate_model_parameter_elasticity}
\alias{calculate_model_parameter_elasticity}
\title{Calculate elasticity based on underlying model parameters}
\usage{
calculate_model_parameter_elasticity(
  param_set,
  class_names,
  transition_matrix,
  fecondity_matrix,
  stoichiometry_array,
  element_names,
  L_max,
  delta_t,
  theta,
  class_lim,
  sexratio,
  gravid,
  fertil,
  growth_rate_coef,
  growth_rate_intercept,
  molt_cycle_a,
  molt_cycle_b,
  molt_cycle_c,
  molt_cycle_d,
  parameter_name
)
}
\arguments{
\item{param_set}{A single parameter set}

\item{class_names}{Vector of size class names}

\item{transition_matrix}{Original growth transition matrix}

\item{fecondity_matrix}{Original fecundity matrix}

\item{stoichiometry_array}{Matrix with stoichiometric data}

\item{element_names}{Vector of element names}

\item{L_max}{Maximum size for growth model}

\item{delta_t}{Time step in days}

\item{theta}{Current temperature in Celsius}

\item{class_lim}{Vector of size class limits}

\item{sexratio}{Sex ratio (proportion of females)}

\item{gravid}{Proportion of gravid females}

\item{fertil}{Vector of fertility rates by size class}

\item{growth_rate_coef}{Growth rate coefficient in growth model}

\item{growth_rate_intercept}{Growth rate intercept in growth model}

\item{molt_cycle_a}{Parameter beta_1 in molt cycle equation}

\item{molt_cycle_b}{Parameter alpha_1 in molt cycle equation}

\item{molt_cycle_c}{Parameter beta_2 in molt cycle equation}

\item{molt_cycle_d}{Parameter alpha_2 in molt cycle equation}

\item{parameter_name}{Name of the parameter to modify}
}
\value{
A data.table with elasticity results
}
\description{
This function calculates elasticity by modifying the underlying model parameters
that determine the growth and fecundity matrices, then measuring the effect on lambda and phosphorus.
}
