% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_dynamics.R
\name{calculate_transition_rates}
\alias{calculate_transition_rates}
\title{Calculate transition rates for a range of temperatures}
\usage{
calculate_transition_rates(temp_range, L_max, delta_t, class_lim, class_names)
}
\arguments{
\item{temp_range}{Numeric vector of temperatures to calculate transition rates for}

\item{L_max}{Maximum size in mm}

\item{delta_t}{Time step in days}

\item{class_lim}{Vector of size class limits}

\item{class_names}{Vector of size class names}
}
\value{
A data frame in long format containing transition rates for each temperature
}
\description{
This function calculates transition rates between size classes for a range of temperatures
using the growth_rates_matrix function.
}
