% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{create_model_parameter_elasticity_figure}
\alias{create_model_parameter_elasticity_figure}
\title{Create model parameter elasticity figure}
\usage{
create_model_parameter_elasticity_figure(elasticity_results)
}
\arguments{
\item{elasticity_results}{Data table with model parameter elasticity results}
}
\value{
A ggplot object for the model parameter elasticity figure
}
\description{
This function creates a figure showing the sensitivity of population growth rate
and phosphorus content to changes in underlying model parameters.
}
