% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gammarus-phosphorum-package.R
\docType{package}
\name{gammarus.phosphorum-package}
\alias{gammarus.phosphorum}
\alias{gammarus.phosphorum-package}
\title{gammarus.phosphorum: Population-level Tests of the Growth Rate Hypothesis in Gammarus fossarum}
\description{
A matrix-based framework for testing ecological stoichiometry theories at the population level. Integrates phosphorus content measurements with a size-structured population model for Gammarus fossarum to investigate whether the Growth Rate Hypothesis applies at the population level. Includes functions for matrix population modeling, stoichiometric calculations, and comprehensive simulation tools to explore the relationship between population growth rate and elemental composition.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/titouan-dionet/gammarus-phosphorum}
  \item Report bugs at \url{https://github.com/titouan-dionet/gammarus-phosphorum/issues}
}

}
\author{
\strong{Maintainer}: Titouan Dionet \email{titouan.dionet@univ-lorraine.fr} (\href{https://orcid.org/0009-0006-5602-1873}{ORCID}) [copyright holder]

}
\keyword{internal}
