% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{generate_multi_param_combinations}
\alias{generate_multi_param_combinations}
\title{Generate random parameter sets for multi-parameter simulations}
\usage{
generate_multi_param_combinations(
  temps,
  n_iter = 10000,
  min_val = 0.001,
  max_val = 1
)
}
\arguments{
\item{temps}{Vector of temperatures to simulate}

\item{n_iter}{Number of random parameter sets per temperature}

\item{min_val}{Minimum parameter value}

\item{max_val}{Maximum parameter value}
}
\value{
A data.table with random parameter sets
}
\description{
This function generates random parameter sets for multi-parameter simulations
where all parameters vary simultaneously
}
