% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{generate_single_param_combinations}
\alias{generate_single_param_combinations}
\title{Generate parameter combinations for single parameter simulations}
\usage{
generate_single_param_combinations(
  temps,
  classes,
  base_surv_rates,
  n_iter = 1000
)
}
\arguments{
\item{temps}{Vector of temperatures to simulate}

\item{classes}{Vector of size classes}

\item{base_surv_rates}{Vector of base survival rates}

\item{n_iter}{Number of iterations (parameter values) per combination}
}
\value{
A data.table with parameter combinations
}
\description{
This function generates parameter combinations for single parameter simulations
where one parameter varies while others stay constant
}
