% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_dynamics.R
\name{growth_rates_matrix}
\alias{growth_rates_matrix}
\title{Population growth rates matrix determination}
\usage{
growth_rates_matrix(
  N_ind = 10000,
  L_max,
  delta_t,
  theta,
  class_lim,
  class_names = NULL,
  growth_rate_coef = 0.0014,
  growth_rate_intercept = -0.0024
)
}
\arguments{
\item{N_ind}{Integer. Total number of individuals in each class needed for determining
the transition rate (default: 10,000).}

\item{L_max}{Numeric. Maximum size of individuals for the logistic growth model.}

\item{delta_t}{Numeric. Time step in days.}

\item{theta}{Numeric. Target temperature in Celsius.}

\item{class_lim}{Numeric vector. Boundaries of the different classes; note that the upper limit
must be specified.}

\item{class_names}{Character vector. Names of the different classes (default: \code{NULL}).}

\item{growth_rate_coef}{Numeric. Coefficient for temperature in growth rate equation (default: 0.0014).}

\item{growth_rate_intercept}{Numeric. Intercept in growth rate equation (default: -0.0024).}
}
\value{
A data frame representing the transition matrix for the given parameters, with rows and
columns named according to class_names if provided.
}
\description{
This function determines the transition matrix from one size class to another for individuals
of a population of gammarus whose growth is governed by a temperature-dependent logistic model.
}
\details{
The function calculates transition probabilities between size classes based on a logistic growth model
where the growth rate depends on temperature according to the equation: r(θ) = 0.0014\emph{θ - 0.0024.
The size of individuals after the time step delta_t is calculated using the logistic equation:
L(t) = L_max / (1 + ((L_max/L_init) - 1) * exp(-r}t))
}
\examples{
\dontrun{
# Calculate transition matrix for gammarus with 5 size classes
class_lim <- c(1.5, 3.5, 5.2, 6, 7, 11)
class_names <- c("J1", "J2", "A1", "A2", "A3")
trans_mat <- growth_rates_matrix(
  L_max = 8.5, 
  delta_t = 30, 
  theta = 12, 
  class_lim = class_lim, 
  class_names = class_names
)
}

}
