% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{run_single_simulation}
\alias{run_single_simulation}
\title{Run a single population simulation}
\usage{
run_single_simulation(
  theta,
  surv_rates,
  transition_matrix,
  fecondity_matrix,
  class_names,
  stoichiometry_array,
  element_names
)
}
\arguments{
\item{theta}{Temperature in Celsius}

\item{surv_rates}{Vector of survival rates for each size class}

\item{transition_matrix}{Growth transition matrix}

\item{fecondity_matrix}{Fecundity matrix}

\item{class_names}{Vector of size class names}

\item{stoichiometry_array}{Matrix with stoichiometric data}

\item{element_names}{Vector of element names}
}
\value{
A data.table with simulation results
}
\description{
This function runs a single population dynamics simulation with the given parameters
and returns population metrics (growth rate, SSD, phosphorus content)
}
