% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{sample_elasticity_parameters}
\alias{sample_elasticity_parameters}
\title{Sample parameter sets for elasticity analysis}
\usage{
sample_elasticity_parameters(multi_param_results, n_samples = 10000)
}
\arguments{
\item{multi_param_results}{Results from multi-parameter simulations}

\item{n_samples}{Number of parameter sets to sample}
}
\value{
A data.table with sampled parameter sets
}
\description{
This function samples parameter sets from multi-parameter results for elasticity analysis
}
