% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{save_figure}
\alias{save_figure}
\title{Save a figure in multiple formats}
\usage{
save_figure(
  plot,
  basename,
  dir,
  width,
  height,
  units = "in",
  dpi = 300,
  formats = c("png", "svg")
)
}
\arguments{
\item{plot}{The ggplot object to save}

\item{basename}{Base filename without extension}

\item{dir}{Directory where the files should be saved}

\item{width}{Width of the figure (in the specified units)}

\item{height}{Height of the figure (in the specified units)}

\item{units}{Units for width and height ('in', 'cm', 'mm', or 'px')}

\item{dpi}{Resolution in dots per inch}

\item{formats}{Vector of formats to save (default: c("png", "svg"))}
}
\value{
Path to the last saved file
}
\description{
Saves a ggplot object as both PNG and SVG files with the specified dimensions
}
