% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{calculate_elasticity}
\alias{calculate_elasticity}
\title{Calculate elasticity for different demographic parameters}
\usage{
calculate_elasticity(
  param_set,
  class_names,
  transition_matrix,
  fecondity_matrix,
  stoichiometry_array,
  element_names,
  parameter_type = "survival"
)
}
\arguments{
\item{param_set}{A single parameter set}

\item{class_names}{Vector of size class names}

\item{transition_matrix}{Growth transition matrix}

\item{fecondity_matrix}{Fecundity matrix}

\item{stoichiometry_array}{Matrix with stoichiometric data}

\item{element_names}{Vector of element names}

\item{parameter_type}{Type of parameter to modify: "survival", "fecundity", or "growth"}
}
\value{
A data.table with elasticity results
}
\description{
This function calculates elasticity for a given parameter set by reducing
survival, fecundity, or growth rates by 10\% and measuring the effect on lambda and phosphorus
}
