% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{create_elasticity_figure}
\alias{create_elasticity_figure}
\title{Create comprehensive elasticity figure}
\usage{
create_elasticity_figure(elasticity_results, analysis_type)
}
\arguments{
\item{elasticity_results}{Data table with comprehensive elasticity analysis results.}

\item{analysis_type}{Type of analysis to coduct: \code{"survival"}, \code{"fecundity"} or \code{"growth"}.}
}
\value{
A ggplot object for Figure 3, S3 or S4
}
\description{
This function creates Figure 3, S3 and S4 for the manuscript showing
the sensitivity of population growth rate and phosphorus content
to survival, fecundity, and growth rates.
}
