% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gammarus-phosphorum-package.R
\docType{package}
\name{gammarus.phosphorum-package}
\alias{gammarus.phosphorum}
\alias{gammarus.phosphorum-package}
\title{gammarus.phosphorum: Population-level Tests of the Growth Rate Hypothesis in Gammarus fossarum}
\description{
Companion code to Dionet et al. (in prep.) testing the Growth Rate Hypothesis at the population level in Gammarus fossarum. Provides functions for size-structured matrix population modeling, stoichiometric calculations, and simulation tools to explore the relationship between population growth rate and phosphorus content. Raw data are archived separately [DOI to be completed]. The manuscript is available as a preprint [DOI to be completed].
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/titouan-dionet/gammarus-phosphorum}
  \item \url{[DOI to be completed]}
  \item Report bugs at \url{https://github.com/titouan-dionet/gammarus-phosphorum/issues}
}

}
\author{
\strong{Maintainer}: Titouan Dionet \email{titouan.dionet@univ-lorraine.fr} (\href{https://orcid.org/0009-0006-5602-1873}{ORCID}) [copyright holder]

}
\keyword{internal}
